requirements.txt
flask
pillow 
